page_analyzer/pyproject.toml
[project]
name = "page_analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "gunicorn>=23.0.0",
    "Flask>=3.1.0",
    "python-dotenv==1.0.1",
    "psycopg2-binary==2.9.10",
    "validators==0.34.0"
]

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]
